+++ src/components/Header/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import {Header} from '../../stories/Header'
+
+
+function HeaderComponent() {
+
+    return (
+        <Header/>
+    )
+
+}
+
+
+export default HeaderComponent