+++ CHANGELOG.md
@@ -4,11 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
 
-[Unreleased]: https://github.com/tognia/shiny-agency/tree/master
+## [1.1.0] - 2021-12-31
+### Added
+ - Add a task with date
+ - Edit a task
+ - Archive a task 
+ - Restore a task
+ - Filter the tasks 
 
 
+## [1.1.0] - 2021-12-31
+### Added
+- Select Task on the List 
+- Edit Task. 
+- Delete Task. 
+
+## [1.0.0] - 2021-12-14
+### Added
+- Add a new task to the taskList
+- Show the TaskList
+- Source of datas added.
 
 ## [0.1.0] - 2021-12-14
 ### Added
@@ -22,8 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [Unreleased]: 
-[0.0.1]: https://github.com/ZiyoumaSchool/todoappgo
+[0.2.0]: https://github.com/ZiyoumaSchool/todoappgo
 ### Added
-- Basics sources of the app 
+- Feature for adding a new task to the TaskList
+- Show the TasList Updated
  
-[0.1.0]:https://github.com/ZiyoumaSchool/todoappgo